refactor: fagLists constants 폴더로 디렉토리 변경
